test(plugin-computeruse): consolidate duplicate scene, platform, and service test suites (#28796)
